added cache and inlined intrinstincs

+++ src/tables.rs
@@ -2,7 +2,7 @@ use crate::layer::Layer;
 use crate::vec::i8x16;
 
 pub struct Tables {
-  pub cur_layer: usize,
+  pub cur_layer: i8,
   pub layers: Vec<Layer>,
   pub pairs: Vec<Vec<usize>>,
   pub goal: i8x16,
